ajout produit terminé et indexEmployé finis (voir si on a le temps pour vérif js)

+++ Expeditor/src/fr/eni_ecole/fr/jee/dal/CommandeDAO.java
@@ -102,7 +102,7 @@ public class CommandeDAO {
 		return liste;
 	}
 	
-	public static Commande findById(int id){
+	public static Commande findById(Long id){
 		
 		EntityManager em = PersistenceManager.createEntityManager();
 		EntityTransaction trans = em.getTransaction();
